refactor: Refactor file structure for best and sale product components
