+++ components/items/render/best-products.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import ShowBestProductData from "../ShowBestProductData.jsx";
+import ShowBestProductData from "../page/ShowBestProductData.jsx";
 
 const root = ReactDOM.createRoot(document.getElementById('best-products-container'));
 root.render(<ShowBestProductData productNumber={4} />);
