+++ index.html
@@ -82,7 +82,7 @@
         /></a>
       </div>
     </footer>
-  <script type="module" src="/components/main/best-products.jsx"></script>
-  <script type="module" src="/components/main/sale-products.jsx"></script>
+  <script type="module" src="/components/items/render/best-products.jsx"></script>
+  <script type="module" src="/components/items/render/sale-products.jsx"></script>
   </body>
 </html>
