+++ public/home.html
@@ -28,7 +28,7 @@
     <!-- Google Analytics 계정을 생성해 ) 본인의 Tracking ID(예: G-XXXXXXXXXX)로 변경이 필요합니다. -->
 <!--    <script-->
 <!--      async-->
-<!--      main="https://www.googletagmanager.com/gtag/js?id=G-본인의_Tracking_ID"-->
+<!--      render="https://www.googletagmanager.com/gtag/js?id=G-본인의_Tracking_ID"-->
 <!--    ></script>-->
 <!--    <script>-->
 <!--      window.dataLayer = window.dataLayer || [];-->
