+++ styles/components/saleProductCard.css
@@ -1,6 +1,6 @@
 .product-image{
-    height: 282px;
-    width: 282px;
+    height: 220px;
+    width: 220px;
     border-radius: 20px;
     display: flex;
     flex-direction: column;
@@ -16,7 +16,7 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    gap: 10px;
+    gap: 6px;
     align-self: stretch;
 }
