+++ styles/global.css
@@ -5,8 +5,8 @@
   - CSS 변수를 사용하면 테마 색상을 중앙에서 관리할 수 있어 사이트 전반에 걸쳐 일관된 스타일을 유지하고, 필요한 경우 쉽게 색상을 변경할 수 있어요.
   - 전역적으로 사용할 수 있도록 root CSS selector(`:root` pseudo-class)에서 선언하는 것이 일반적이에요.
   - 변수명을 등록할 때는 hyphen 두 개(`--`) 뒤에 원하는 이름을 붙여주면 되는데, 이때 색상명이나 색상을 잘 설명하는 별칭을 부여하면 돼요. 
-    (예: 메인 컬러인 하늘색 색상은 `--blue` 또는 `--main-color` 등으로 등록할 수 있겠죠. 
-    포인트 컬러가 다양하거나 등록할 블루 계열 색상이 많아 이런 포괄적인 이름으로 구분이 어려운 경우에는 `--main-button-color` 같이 해당 색상의 사용 용도로 표현하기도 해요.)
+    (예: 메인 컬러인 하늘색 색상은 `--blue` 또는 `--render-color` 등으로 등록할 수 있겠죠.
+    포인트 컬러가 다양하거나 등록할 블루 계열 색상이 많아 이런 포괄적인 이름으로 구분이 어려운 경우에는 `--render-button-color` 같이 해당 색상의 사용 용도로 표현하기도 해요.)
   - 등록된 색상을 적용할 때는 hex code 대신 `var(변수명)`를 넣어주면 됩니다.
 
   - 참고 문서: https://developer.mozilla.org/ko/docs/Web/CSS/Using_CSS_custom_properties
