+++ styles/items.css
@@ -29,4 +29,13 @@
     gap: 40px;
 }
 
+.product-category {
+    color: #111827;
+    font-family: Pretendard, sans-serif;
+    font-size: 20px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 32px;
+}
+
 /* BestProductCard 컴포넌트 스타일 적용 */
